+++ src/RenderObject/Vector2D.java
@@ -1,14 +1,14 @@
 package RenderObject;
 public class Vector2D {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     
     public Vector2D() {
         this.x = 0;
         this.y = 0;
     }
     
-    public Vector2D(int x, int y) {
+    public Vector2D(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -16,13 +16,13 @@ public class Vector2D {
     /** get X coordinate of this vector
     @return return integer of this vector in X axis
     **/
-    public int getXCoord() {
+    public double getXCoord() {
         return this.x;
     }
     /** get X coordinate of this vector
     @return return integer of this vector in Y axis
     **/
-    public int getYCoord() {
+    public double getYCoord() {
         return this.y;
     }
 }
